Using custom queryset for CourseEnrollment model
This queryset shortcuts attempts to count all rows in the table to help with rendering of the admin table.
